Lower default rebuild time
When there are more disks than replicas in the cluster and a disk is removed
(weight set to 0.0) the data is distributed across all remaining disks. Thus it
is not required to replicate all the data to a single disk, but to many disks in
the cluster and each disk get's only a fraction of the data. This is much
faster.

A more exact way to calculate this time will follow soon.